Refactor SingleProgramView to simplify code for displaying program changes
